Add commandline module
This module is used to parse commandline arguments.
It is a module so the app can use it the same way as
any scripts.

The idea is to import commandline, which creates
commandline.parser with a number of generic arguments
(e.g. project directory).
Before calling commandline.parse() each script can add
its specific arguments.