Project: init segment names
When the segment names are not read from a json file they have
to be initialized somehow. This function takes either a list
of segment names or a callable to populate the list.
One provided callable generates the segment names by generating
a range of integer names. These are automatically padded to the
necessary number of digits (or to a number set by an optional
argument). The range can be zero- or one-based (zero is default).