Segment: Improve "deleted" infrastructure
In the first version of the script a segment spawned a Git
subprocess to determine its deleter. This slowed down
considerably because of the number of subprocesses.

Now there's a dictionary in the SegmentGrid object that holds
information about deleted files. This will be generated upon
the first request and then cached.
When a segment is asked for its deleter it forwards the request
to SegmentGrid.deleted_by(voice, segment).