GitRepo.changed_files() and .changed_segments()
Return a list with files that have changed on "another" branch.
First it determines a common ancestor, then which files have
changed between this commit and the other branch's head.

.changed_segments() limits the output to the 'music' path of the
project and returns the segments as a dictionary that contains
segment lists for all affected voices.

Note that this may return flawed output when there are merge
commits in the history, i.e. if the current branch (usually
master) has been merged into the working branch.