gen-voice: Segment properties, read time signature
The content of a segment gets parsed into a music variable
and properties (the properties will later be used to set the
opening of grid cells).

When generating the segment templates and empty segment stubs
we will determine if time signatures (and other properties)
are redundant and only insert them in the music when they are
actually changed.